complete renaming of tracer to stepper
