+++ src/runner/sourceRunner.ts
@@ -6,7 +6,7 @@ import { CSEResultPromise, evaluate as CSEvaluate } from '../cse-machine/interpr
 import { ExceptionError } from '../errors/errors';
 import { RuntimeSourceError } from '../errors/runtimeSourceError';
 import { TimeoutError } from '../errors/timeoutErrors';
-import { getSteps } from '../tracer/steppers';
+import { getSteps } from '../stepper/steppers';
 import { sandboxedEval } from '../transpiler/evalContainer';
 import { transpile } from '../transpiler/transpiler';
 import { toSourceError } from './errors';
